js/bundle: Remove Node-specific library dependency from encoder
Now the wbn module depends only on cborg.
